# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/stays": {
      "get": {
        "operationId": "StaysController_list",
        "parameters": [
          {
            "name": "providerId",
            "required": false,
            "in": "query",
            "description": "Filter by provider ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by listing status",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "PUBLISHED", "PAUSED", "ARCHIVED"]
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "Filter by city",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name (ILIKE)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived listings",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": ["createdAt", "name"]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of stays"
          }
        },
        "summary": "List stay listings",
        "tags": ["stays"]
      }
    },
    "/v1/stays/{id}": {
      "get": {
        "operationId": "StaysController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stay listing detail with amenities"
          }
        },
        "summary": "Get stay listing by id",
        "tags": ["stays"]
      }
    },
    "/v1/stays/{providerId}/listings": {
      "post": {
        "operationId": "StaysController_createStayListing",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stay created in DRAFT status"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create stay listing",
        "tags": ["stays"]
      }
    },
...
  },
  "info": {
    "title": "TripAvail API",
    "description": "Two-sided travel marketplace API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateListingDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Beachfront Villa"
          },
          "address": {
            "type": "string",
            "example": "123 Ocean Drive"
          },
          "city": {
            "type": "string",
            "example": "Miami"
          },
          "latitude": {
            "type": "number",
            "example": 25.790654
          },
          "longitude": {
            "type": "number",
            "example": -80.1300455
          },
          "description": {
            "type": "string",
            "example": "Spacious villa with ocean views"
          },
          "checkInTime": {
            "type": "string",
            "example": "14:00"
          },
          "checkOutTime": {
            "type": "string",
            "example": "11:00"
          },
          "amenityIds": {
            "example": ["amenity_cuid1", "amenity_cuid2"],
            "description": "Amenity IDs to attach",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "address",
          "city",
          "latitude",
          "longitude",
          "description",
          "checkInTime",
          "checkOutTime"
        ]
      },
...
    }
  }
}